Max company descriptions a max of 50 characters.

+++ app/Http/Controllers/CompaniesController.php
@@ -87,9 +87,11 @@ class CompaniesController extends Controller
           $company->save();
         } elseif (!$request->has('image')) {
           $image = $company->image;
-          $company->image_id = null;
-          $company->save();
-          $image->delete();
+          if ($image) {
+              $company->image_id = null;
+              $company->save();
+              $image->delete();
+          }
         }
 
         return response()->json([
